+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 3.0.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Periodic export job.
+  [folix-01]
 
 
 3.0.2 (2024-05-05)